Abstract get_secrets function
